Added missing variable 'print-weights' to the XfstCompiler constructor with impl argument. Added support for prolog format to hfst-fst2txt and a simple prolog test. Fixed a typo in configure.ac. Removed unnecessary commented text from HfstTransitionGraph.h.
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@3554 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/fst2txt-functionality.sh
@@ -30,6 +30,9 @@ if ((test -z "$i") || $TOOLDIR/hfst-format --list-formats | grep $i > /dev/null)
         if ! $TOOLDIR/hfst-fst2txt -f pckimmo < cat$i > test.txt ; then
             exit 1
         fi
+        if ! $TOOLDIR/hfst-fst2txt -f prolog < cat$i > test.txt ; then
+            exit 1
+        fi
         rm test.txt
     fi
 fi
